bot/pyproject.toml
[project]
name = "durian-invest"
version = "0.1.0"
description = "Simple Trading Bot Testing"
requires-python = ">=3.10"
dependencies = [
    "alpaca-py>=0.30.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.2.0",
    "apscheduler>=3.10.0",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "google-search-results>=2.4.2",
    "supabase>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
